Split postgres implementation of database store in its own file
