+++ core/src/lib.rs
@@ -118,6 +118,7 @@ pub mod deno {
 }
 
 pub mod databases_store {
+    pub mod postgres;
     pub mod store;
 }
